backend-python/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
wallesmart-sdk[fastapi]==1.0.0
